Fix not being able to delete folder in data-model (#11432)
* fix not being able to delete folder in data-model

* add e2e tests

* fix e2e tests

* fix e2e tests

+++ api/src/services/collections.ts
@@ -377,7 +377,9 @@ export class CollectionsService {
 		}
 
 		await this.knex.transaction(async (trx) => {
-			await trx.schema.dropTable(collectionKey);
+			if (collectionToBeDeleted!.schema) {
+				await trx.schema.dropTable(collectionKey);
+			}
 
 			// Make sure this collection isn't used as a group in any other collections
 			await trx('directus_collections').update({ group: null }).where({ group: collectionKey });
